feat(back, order): db backed statuses (#2)
-Rework order statuses
-Modify dto parameter names to "request" for better readability
-Rework entities for better handling of validation and db annotations
-Max delivery attempts config prop moved to db 

-feat: total price calculation
-feat: handle cancelled/returned order stock

+++ src/main/java/com/santiGalarza/order_management/product/CreateProductRequest.java
@@ -37,5 +37,5 @@ public class CreateProductRequest {
 
     @NotNull
     @Positive
-    private Float weight;
+    private BigDecimal weight;
 }
